+++ AdventOfCode/Challenges/Challenge2b.cs
@@ -17,7 +17,7 @@ namespace AdventOfCode.Challenges
 
 		public async Task<string> RunAsync()
 		{
-			int[] program = (await File.ReadAllTextAsync("Assets/Challenge2.txt")).Split(',').Select(s => int.Parse(s)).ToArray();
+			long[] program = (await File.ReadAllTextAsync("Assets/Challenge2.txt")).Split(',').Select(s => long.Parse(s)).ToArray();
 
 			for (int noun = 0; noun < 100; noun++)
 			{
@@ -26,7 +26,7 @@ namespace AdventOfCode.Challenges
 					program[1] = noun;
 					program[2] = verb;
 					var computer = new SimpleRunner(program);
-					int result = computer.Execute();
+					int result = (int)computer.Execute();
 					if (result == 19690720)
 					{
 						return (100 * noun + verb).ToString();
